+++ src/main/java/com/example/ecommercebe/repositories/ShoppingCartRepository.java
@@ -4,5 +4,5 @@ import com.example.ecommercebe.entities.ShoppingCart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ShoppingCartRepository extends JpaRepository<ShoppingCart, Integer> {
-    ShoppingCart findShoppingCartByUser_Id(Integer id);
+    ShoppingCart findShoppingCartByUser_Id(Long id);
 }
